change valdators to be scoped .

+++ src/MinimalApi/Services/IPersonService.cs
@@ -6,10 +6,10 @@ namespace MinimumApi.Services
     public interface IPersonService
     {
         Task<IEnumerable<Person>> GetAllPeopleAsync();
-        Task<Person> GetPersonByIdAsync(int id);
+        Task<Person> GetPersonByIdAsync(long id);
         Task<Person> GetPersonByNameAsync(string name);
         Task AddPersonAsync(Person person);
         Task UpdatePersonAsync(Person person);
-        Task DeletePersonAsync(int id);
+        Task DeletePersonAsync(long id);
     }
 }
